+++ components/playground/instance/tidb.go
@@ -28,12 +28,13 @@ type TiDBInstance struct {
 	instance
 	pds []*PDInstance
 	Process
-	enableBinlog bool
-	isDisaggMode bool
+	tiproxyCertDir string
+	enableBinlog   bool
+	isCSEMode      bool
 }
 
 // NewTiDBInstance return a TiDBInstance
-func NewTiDBInstance(binPath string, dir, host, configPath string, id, port int, pds []*PDInstance, enableBinlog bool, isDisaggMode bool) *TiDBInstance {
+func NewTiDBInstance(binPath string, dir, host, configPath string, id, port int, pds []*PDInstance, tiproxyCertDir string, enableBinlog bool, isCSEMode bool) *TiDBInstance {
 	if port <= 0 {
 		port = 4000
 	}
@@ -47,14 +48,15 @@ func NewTiDBInstance(binPath string, dir, host, configPath string, id, port int,
 			StatusPort: utils.MustGetFreePort("0.0.0.0", 10080),
 			ConfigPath: configPath,
 		},
-		pds:          pds,
-		enableBinlog: enableBinlog,
-		isDisaggMode: isDisaggMode,
+		tiproxyCertDir: tiproxyCertDir,
+		pds:            pds,
+		enableBinlog:   enableBinlog,
+		isCSEMode:      isCSEMode,
 	}
 }
 
 // Start calls set inst.cmd and Start
-func (inst *TiDBInstance) Start(ctx context.Context) error {
+func (inst *TiDBInstance) Start(ctx context.Context, _ utils.Version) error {
 	configPath := filepath.Join(inst.Dir, "tidb.toml")
 	if err := prepareConfig(
 		configPath,